send status in service history api

+++ app/Http/Controllers/Admin/ServiceCenterController.php
@@ -182,8 +182,10 @@ class ServiceCenterController extends Controller
 
         if(count($file)>0) {
             //Delete previous image from folder
-            unlink($model->file);
-
+            if(($model->file)){
+                unlink($model->file);
+            }
+            
             // Files destination
             $destinationPath = 'public/uploads/service_center/';
